+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXGJ24PZ2K5MQZ14KHZYHVET"
+        "01KXGQ6BPCRQHFMZ8JDQJS4JDP"
     )
     r"""The unique identifier of the data source"""