+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-14T14:56:04.781Z"
+    created: Optional[str] = "2026-07-14T16:25:32.078Z"
 
-    updated: Optional[str] = "2026-07-14T14:56:04.781Z"
+    updated: Optional[str] = "2026-07-14T16:25:32.078Z"
 
     updated_by_id: OptionalNullable[str] = UNSET