+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -40,9 +40,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-14T14:56:04.781Z"
+    created: Optional[str] = "2026-07-14T16:25:32.078Z"
 
-    updated: Optional[str] = "2026-07-14T14:56:04.781Z"
+    updated: Optional[str] = "2026-07-14T16:25:32.078Z"
 
     updated_by_id: OptionalNullable[str] = UNSET