+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXGJ24PYHA5B5SE3RDNSP09Y"
+        "01KXGQ6BPASGJ39CRBC4BT3VMQ"
     )
     r"""The unique identifier of the data source"""