+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXGJ24Q2BRD2QHQ1ZP8G8STH"
+        "01KXGQ6BPH7JNBQ7G6Z6B3V6GP"
     )
     r"""The unique identifier of the data source"""