+++ packages/orq-rc/RELEASES.md
@@ -9459,3 +9459,13 @@ Based on:
 - [python v4.12.0-rc.24] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.24 - packages/orq-rc
+
+## 2026-07-14 16:28:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.25 - packages/orq-rc